REFACTOR: InsertModeSelector to TS
