+++ packages/neos-ui-constants/src/dndTypes.ts
@@ -1,5 +1,6 @@
 const dndTypes = {
     NODE: 'neos-tree-node',
     MULTISELECT: 'neos-multiselect-value'
-};
+} as const;
+
 export default dndTypes;
